requirements.txt
absl-py==0.13.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
arrow==1.0.3
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
binaryornot==0.4.4
black==21.5b1
bleach==3.3.0
bokeh==2.3.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colored==1.4.2
cookiecutter==1.7.2
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
flatbuffers==1.12
gast==0.4.0
Gooey==1.0.8.1
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.0.8
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
livelossplot==0.5.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
pygtrie==2.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark==3.1.1
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transformers==4.7.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wxPython==4.1.1
